checkpoint

+++ src/Z80.h
@@ -142,7 +142,7 @@ private:
   void decode();
   AluOut exec(uint8_t src) const;
   uint16_t reg_fetch() const;
-  uint16_t reg_fetch8() const;
+  uint8_t reg_fetch8() const;
   uint16_t reg_fetch16() const;
   void reg_put(int mux, uint8_t reg);
 };
